Change shell tool structure
Interface to shell tools will be via dnsxd-rb and dnsxd-admin scripts.
These scripts are verbatim copies of 'rel/files/dnsxdshell_tmpl' which
is a simple wrapper that swaps dashes in it's own name for spaces and
then calls the result. For example, if named 'dnsxd-admin' it will call
'dnsxd admin'.