Add dnsxd_couch_log_server
Server holds a named ETS table which other processes write to via
dnsxd_log/1. This ETS table is periodically renamed and a new one
created to take it's place. The contents of the renamed table are
then merged into one of a series of docs which together cover a
rolling window of time. The default number of documents is 72
with each document covering a period of 1200 seconds. This can be
changed by modifying the document called dnsxd_couch_log_config
which will be created automatically if it doesn't exist.

Todo:
* Monitor for settings doc changes
* Monitor for conflicted log docs
* Integrate in supervisor tree